pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-news-verifier"
version = "0.1.0"
description = "Daily AI podcast verifier with LINE notifications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "uvicorn>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
